requirements.txt
flask
matplotlib
langdetect
translate
PyPDF2
python-docx

